a bit of reorganization
