+++ src/route/user/wcl.ts
@@ -9,7 +9,6 @@ import {
 } from "passport-oauth2";
 import * as cache from "../../cache.ts";
 import * as Sentry from "@sentry/node";
-import axios from "axios";
 import * as crypto from "node:crypto";
 
 const userInfoQuery = gql`
@@ -69,41 +68,7 @@ async function fetchWclProfile(
   return parseWclProfile(profile);
 }
 
-export async function refreshWclToken(
-  refreshToken: string
-): Promise<string | undefined> {
-  const basicAuth = Buffer.from(
-    `${process.env.WCL_CLIENT_ID}:${process.env.WCL_CLIENT_SECRET}`
-  ).toString("base64");
-  for (let i = 0; i < 3; i++) {
-    try {
-      const response = await axios.postForm(
-        `https://www.${process.env.WCL_PRIMARY_DOMAIN}/oauth/token`,
-        {
-          grant_type: "client_credentials",
-          code: refreshToken,
-          redirect_uri: process.env.WCL_REDIRECT_URL,
-        },
-        {
-          headers: {
-            Accept: "application/json",
-            Authorization: `Basic ${basicAuth}`,
-          },
-        }
-      );
-      return response.data.access_token;
-    } catch (error) {
-      if (error instanceof Error) {
-        console.log(error.cause, error.message, error.name);
-        /* if (error.response?.status === 400) {
-          throw new Error("Invalid refresh token");
-        } */
-      }
-    }
-  }
-}
-
-export async function refreshTokenKey(refreshToken: string) {
+export async function userRefreshTokenKey(refreshToken: string) {
   const hasher = crypto.createHash("sha256");
   hasher.update(refreshToken);
   return `wcl-token-${hasher.digest("hex")}`;
@@ -153,7 +118,7 @@ const wcl =
           }
 
           await cache
-            .set(await refreshTokenKey(refreshToken), accessToken)
+            .set(await userRefreshTokenKey(refreshToken), accessToken)
             .catch(Sentry.captureException);
 
           const [user, created] = await User.findOrCreate({
